moved URL routing to app.js (issue #71)
