+++ views/user/login.js
@@ -4,17 +4,8 @@ var user = require('../../lib/user');
 var utils = require('../../lib/utils');
 
 
-module.exports = function(server) {
-    // Sample usage:
-    // % curl -X POST 'http://localhost:5000/user/login' -d 'assertion=&audience'
-    server.post({
-        url: '/user/login',
-        swagger: {
-            nickname: 'login',
-            notes: 'Sign in via Persona',
-            summary: 'Login'
-        }
-    }, function(req, res) {
+module.exports.postLogin =
+    function(req, res) {
         var POST = req.params;
 
         var assertion = POST.assertion;
@@ -71,5 +62,4 @@ module.exports = function(server) {
                 });
             }
         );
-    });
-};
+    };
