+++ views/user/profile.js
@@ -4,37 +4,8 @@ var auth = require('../../lib/auth');
 var db = require('../../db');
 var user = require('../../lib/user');
 
-module.exports = function(server) {
-    // Sample usage:
-    // % curl -X PUT 'http://localhost:5000/user/profile?_user=ssatoken'
-    /*
-    Optional params (at least one must be provided to have any effect):
-    ?username=new_name
-    &email=new_address
-    */
-    server.put({
-        url: '/user/profile',
-        swagger: {
-            nickname: 'update-profile',
-            notes: 'Update user profile',
-            summary: 'Update user profile'
-        },
-        validation: {
-            _user: {
-                description: 'User (ID or username slug)',
-                isRequired: true
-            },
-            email: {
-                description: 'New user email',
-                isEmail: true,
-                isRequired: false,
-            },
-            username: {
-                description: 'New username',
-                isRequired: false,
-            }
-        }
-    }, db.redisView(function(client, done, req, res) {
+module.exports.putUpdateProfile =
+    db.redisView(function(client, done, req, res) {
         var DATA = req.params;
         var email = req._email;
 
@@ -56,5 +27,4 @@ module.exports = function(server) {
             });
         });
         
-    }));
-};
+    });
