+++ views/user/search.js
@@ -3,35 +3,8 @@ var db = require('../../db');
 var user = require('../../lib/user');
 
 
-module.exports = function(server) {
-
-    // Sample usage:
-    // % curl 'http://localhost:5000/user/search?_user=ssatoken'
-    server.get({
-        url: '/user/search',
-        validation: {
-            _user: {
-                description: "A user's SSA token",
-                isRequired: true
-            },
-            email: {
-                description: 'Email to search for',
-                isRequired: false
-            },
-            id: {
-                description: 'User ID to search for',
-                isRequired: false
-            },
-            devSlug: {
-                description: 'Company Slug to search for',
-                isRequired: false
-            },
-            q: {
-                description: 'Email/user ID/username to search for',
-                isRequired: false
-            }
-        }
-    }, db.redisView(function(client, done, req, res) {
+module.exports.getUser =
+    db.redisView(function(client, done, req, res) {
         var DATA = req.params;
         console.log(DATA);
 
@@ -96,5 +69,4 @@ module.exports = function(server) {
                 done();
             });
         }
-    }));
-};
+    });
